requirements.txt
streamlit
pandas
scikit-learn
seaborn
matplotlib
google-generativeai
numpy
plotly